Added 'Berbanya to Armenya' Location
Editor
- Added tags 'Night Sprite' and 'Night Light' to track night specific
objects.
- Polished

Code
- Added a method scaleOriginal() in MatchTransform.cs which scales the
original scale but maitnains its looking direction as opposed to other
scale methods (scaleAbsolute() - scales the object to the basis but xyz
is absolute; scaleCompletely() - directly copies the scale of the basis)
- Added boolean saveLookDirection to ObjectData.cs to enable/disable
saving of x scale of an object. This syncs with a boolean in
ObjectDataReferences.cs
- Enables or disables night objects with the use of tags in
changeTimeOfDay() method in LevelManager.cs
- Fixed scale in switching characters

Aesthetics
- Added moon in location 'Kwarto ni Ibong Adarna', complete with
lighting.
- Added stars sprites (initially applied lighting as well but removed as
this may cause performance issues). Stars are only plain white circles
currently.