Added Locations & Improved Code
Locations Added
- Forest to Mountain (with sky, stars and moon)
- Mountaintop

Code
- Added a code to enable objectInteraction and npcInteraction collders
while the player is in dialogue and disable the colliders when not in
dialogue.
- Added a 'Zoom()' methods (different versions) in CameraController.cs
that starts a coroutine to zoom the camera (previously done in the
update function which made it difficult to zoom on other classes).
- Added 'CameraTrigger.cs' which calls the 'Zoom' function when a
collider is triggered - can be used for differentiating camera size
among portions of a location.

Misc
- Renamed 'Initial Scene.unity' to '() Initial Scene.unity' to make it
first to appear in the in engine explorer.