Added a script for closing exits
Heading
- Added a global class (DoorAndExitController) that contains the actual
logic of setting up exit data. (Includes methods for closing/opening
exits on other scenes as well as in the current scene)
- Initializations in DoorAndExitController.cs is done every Start() in
LevelManager (through a custom Init()).
- Added 'link' methods in Utilities.cs for opening/closing exits
(overload 1: Accepts 'exit/door' gameObject as a parameter, overload 2:
Accepts string exitName as a parameter, overload 3: Accepts string
exitName as well as string sceneName).

Code
- Improved speech bubble display logic (now has a persisitent coroutine
which checks if the npc is the current objective and is not in a
dialogue in order to display - once it is already displayed it will not
attempt to redesplay.)
- Added codes for fading foregrounds when camera is zoomed; *Commented
out as it does not look good.

Aesthetics
- Created a particle system for alitaptap; added to prefabs.