Added collection system
Heading^
- Complete with UI and methods
- Important methods include:

AddToCollect() - which adds something to collect. Accepts string Name of
the object as well as string setName (determines which set of objects
the item belongs; i.e. bottles, banga; same types). Also accepts a
Sprite icon, which specifies the icon shown in the counterUI.
Collect() - is called in ObjectCollect.cs's update to collect items when
it is allowed. Reduces the count of a specific set and updates the text
UI of the counter.

- Added advanced code for waiting until animation is done before
allowing the player to move.
- Created advanced logic for destroying the item collected when the
itemToGIve of the player is enabled in the animation.