Improved NPCInteraction.cs logic
Heading ^
- Improved NPCInteraction.cs by adding a coroutine to move the player on
the edge of the circle collider of the npc when interaction.
(specialized coroutine used as moveCharacter does not end up to the
exact position)

*Might improve the moveCharacter logic later on

Codes
- Added DeadEndHandler.cs which stops the movement of the player when
collided to a dead end. This is done by starting a coroutine when the
player collided with the dead end. This coroutine only restores movement
when a key is pressed to move to the opposite direction (coroutine is
used for efficiency).
- Added a .NET library called LitJson to read json efficiently. Might
use built in json utility in unity as this .NET library can't directly
parse directly.

Fixes
- MatchTransform.cs - changed Start() to Awake() as there are instances
wherein LevelManager.cs Start() executes before MatchTransform
initializations happen.