Added Barebones Script for Extra Features
Heading^
- Added an assessment manager which is located at the scene '()
Assessment', it basically sends a message to a fungus flowchart to load
the assessment for the current part.
- Added a new coroutine in the LevelLoader.cs which prints 'Loading' in
the unity console while the next scene is loading in the background (add
codes to enable a certain UI element to display while scene is not
finished loading; )

Codes
- Added an action: 'LoadAssessment' as an attribute of an objective,
this triggers the loading of the assessment.
- Added objectives for going to the next part (triggered at the end of
an assessment through objective mapper)

MISC
- Added a scene for assessments (no art yet)
- Added objectives
- Added package 'Unity Standard Assets' which features a library of
useful shaders (got the blur shader in here)