Added a LOT
Heading^
- Saving and loading
*does not save followers, add later
*allow different save file for each part

UI
- Libro ng Kaliwanagan - Polished (added more logi to avoid overlapping
with the generic pause screen)
- LayuninUIManager.cs which handles instructions for minigames (appears
only the first time the player encountered the minigame)
* The appearance of the mini counterpart can be disabled
- Added generic pause menu

Script
- Added a class caled 'GenericMinigameManager' which handles the
launching of LayuninUI by providing minigame specific instructions.

Revisions
- Revised logic of start dialogue to hide objective panel whenever there
is no need to display a text
- Revised libro notification icon as originally it shows a explamation
point when there is a newly activated word, now, it shows the exact
number of words not checked out.
- Revised Dayap minigame UI positioning and added cosmetics previously
added to the original Pedras Platas Forest