Added New UIs and Functionality
UI
- Added confirmation box
- Added interaction prompt (E,W and S)
- Location Marker
- Added gabay panel

Code
- Added a way to trigger a dialogue when an exit is closed (dialog is in
a global flowchart which is randomly called). The character speaking is
also dynamic. It will depend on the current character
- System to display unique sections. and checking if the player is
already in that section ; if not the new section will be displayed,
otherwise nothing will be displayed.

Bugs and Fixes
- Fixed bug of pause menu overriding other menu pauses. (fixed by
determining if another UI element triggers the pause)
- Fixed bug of floating character when a closed door/exit is triggered.

MISC
- Added tutorials each panel with unique icon. (almost brute force
solution)