Added Minigames for Part 9
Heading^
- Added Maria Blanca in castle building minigame
- Added calling of "Lose" block when the player loses a minigame
(calling of method is in specific for each minigame)
- Added walls and better doors graphics in Berbanya
- Added variable "losingPrompt" in GenericMinigameManager.cs to enable
custom prompts for different minigames

- Added actual sprites for wrong choices in (Minigame) Build a Castle.
(Minigame Build a Castle is Complete)
- Created (Minigame) Move Mountain minigame; there is a flow , but
objective mapper is not yet added to a game object since
there are still no objective up to this area; there are no dialogues
please add (there is a readily made flowchart for wrong and correct
placing reactions, please fill;
there is no loading of scene to the scene planned to be loaded after
this minigame.)
- ^Complete synchronization with GenericMinigameManager
- ^ note that castle can only be moved to the X axis while the mountain
can move x and y, and will scale according to its y position, giving the
illusion of it getting closer and closer

Aesthetics
- adjusted font adaptation size minimum of mini LayuninUI as the recent
minigame descriptions does not fit.
- Added a particle prefab for magic effect, this may or may not be
changed to better fx.