Added Codes for Underwater Minigame
Heading^
- Added UnderwaterManager.cs which handles the calling of fungus blocks
to enable flow.
- Like all the of the newer minigames, it waits for the LayuninUI to be
completely faded before starting.
- Created new uniformity among newer minigames (calls blocks rather than
hardcode everything, this can also enable dialogues when sub objectives
are met in minigame)
- Added a temp timer ui

Code
- Added a BetterTimer.cs
- BetterTimer vs Timer - BetterTimer counts down in a coroutine for
efficiency for it not to run after it has not on going, Timer counts
down in Update(); BetterTimer can be paused, Timer cannot.

Bug Fixes
- Fixed 'bug' of disappearing items held by Maria (Happens as the item
is disabled in the animation)
- Fixed 'bug' of other characters not matching scale with player in
'Kwarto ni Maria' location; happens becuase Start() in LevelManager
stops right after matching the scales. This error is encountered becuase
there is a null Door in the list of doors in LevelManager
- Fixed 'bug' of non saving item of player when disableItem() is used;
this happens as disable item regardless of the usage in a character will
empty the item held by the player.