Bug Fixes
Heading^ Bug fixes
- Freezing loading creen: happens because even before the loading screen
completes, another overlaps, since the trigger is set in the position of
the player. Solution; Stopped the coroutine when launching a scene
- Interact prompt appearing. Solution; disabled it when the level is
loaded. (Not sure if it solves the problem)
- Not moving player when scene is loaded through the list. Soultion;
added a boolean to LoadingScreenManager to check if it is used or not.
If it is used, movement is allowed after it completes loading otherwise,
movement is allowed right away

MISC
- Changed path of Talasalitaan.json to persistent data path (same path
to save data); this is done to make
Talasalitaan.json rely on user's own data rather than that present in
the project files/program files. As done before,
when Talasalitaan.json is not found, a new file will be written coping
what is inside Talasalitaan_Backup.json present inthe Project files.